requirement.txt
git+https://github.com/eriksandstroem/evaluate_3d_reconstruction_lib.git
torch-scatter -f https://data.pyg.org/whl/torch-2.1.2+cu118.html
# thirdparty/simple-knn
# thirdparty/diff-gaussian-rasterization
numpy==1.26.4
setuptools==69.5.1
wheel
torchmetrics
pyrender
imgviz
timm
open3d
evo
munch
plyfile
rich
glfw
glm
scipy
tqdm
matplotlib
pyyaml
lightning
ipdb

# torch
# torchvision
numpy==1.26.4
roma==1.5.0
gradio==4.43.0
opencv-python==4.10.0.84
einops==0.8.0
trimesh==4.4.9
tensorboard==2.17.1
pyglet==1.5.29
huggingface-hub[torch]==0.24.6
pillow==10.3.0
gdown==5.2.0
imageio[ffmpeg]==2.35.1
lietorch